Check consistency of restarts in 3d, find and fix bug with periodic dim

+++ m2tools/sim_runner.py
@@ -8,7 +8,7 @@ import checkpoint
 
 class SimulationRunner(object):
 
-    def __init__(self, problem, usempi=False, np=1, suppress_output=True):
+    def __init__(self, problem, usempi=False, np=1):
         self._m2exe = 'src/m2'
         self._np = np
         self._usempi = usempi
@@ -21,7 +21,6 @@ class SimulationRunner(object):
                             hdf5_cadence=0.0,
                             tpl_cadence=0.0,
                             model_parameters="")
-        self._suppress_output = suppress_output
 
     def update_args(self, **kwargs):
         self._m2args.update(kwargs)
@@ -41,10 +40,10 @@ class SimulationRunner(object):
                 argstr += ' --{0}={1}'.format(k.replace('_', '-'), v)
         return argstr
 
-    def run(self):
+    def run(self, remove_chkpt=True, suppress_output=True):
         cmd = self.get_command()
         arg = shlex.split(cmd)
-        stdout = subprocess.PIPE if self._suppress_output else None
+        stdout = subprocess.PIPE if suppress_output else None
         p = subprocess.Popen(arg, stdout=stdout)
         p.communicate()
 
@@ -57,6 +56,8 @@ class SimulationRunner(object):
         for f in chkpt.cell_primitive:
             fields[f] = chkpt.cell_primitive[f][...]
         chkpt.close()
-        os.remove('chkpt.final.h5')
+
+        if remove_chkpt:
+            os.remove('chkpt.final.h5')
 
         return fields
